+++ src/synth/keyboard_sdl.ml
@@ -93,7 +93,10 @@ class keyboard ~kind velocity =
     method remaining = -1
     method abort_track = ()
     method self_sync = (`Static, false)
-    method output = if AFrame.is_partial self#memo then self#get_frame self#memo
+
+    method output =
+      if self#is_ready && AFrame.is_partial self#memo then self#get self#memo
+
     val mutable window = None
 
     method wake_up _ =